Reorganized functions to prevent duplicate calls.
The functions that process repos should only be called after all
the repositories have been collected from the github API. Since
we are using a recursive function, that task of processing repos
should be placed in the ending section of the recursive function.

Use of global (or rather wide scope) variables has now been reduced,
and the necessary variables are passed as arguments or returned as
values from the involved functions.