Updated the logic of the groups page.
The function to get the repositories from the cache was moved to the common
ncipnamespace.js file and it is now called from both the index and group html
pages. The logic of the group page was also updated after the model of the
index page, to manage the cases when data has not been modified and when
response from the server is refuse, in which cases the repository information
is obtained from the local storage cache.