Added external repos and fixed IE8 compatibility.
Added the group of repos from the OpenXIP project.  The process involved
updating the ncip-group.json and ncip-repo.json files. Then using in them the
full_name of the repositories instead of only the name. From the full_name we
programmatically gather the list of organizations hosting the repositories, and
then use that list in order to query the Github API.

Fixed the incompatibility with IE8. Alternative code is now used to gather data
from the Github API via the getJSON function call, while the other cases use
direct XMLHttpRequest() calls.

Added the use of a scopeless authorization token. This additionally help to
raise the rate limit of API requests to 5000 per hour, instead of the default
value of 60 per hour.

Tested these modifications in the following browsers:

  *  Firefox 24.0 in Ubuntu Linux 12.10
  *  Chromium 28.0.1500.71 Ubuntu 12.10
  *  Internet Explorer 10.0.9 on Windows 7
  *  Internet Explorer 10.0.9 on Windows 7 - On IE9 Compatibility Mode
  *  Internet Explorer 10.0.9 on Windows 7 - On IE8 Compatibility Mode
  *  Chromium 29.0.1547  on Windows 7
  *  Firefox 22.0 on Windows 7
  *  Safari on iPad (iOS 6.1.3)
  *  Chromium 29.0.1547  on ChromeOS (Chromebook)
  *  Chrome 28 on Android 4.3 (Nexus 4 phone)
  *  Firefox 24 on Android 4.3 (Nexus 4 phone)